🚚 hook 함수 로 변경
sub-issues:
- [NAILCASE-405](https://nailcase.atlassian.net/browse/NAILCASE-405)
